refactor: Improve article URL handling in Grab component
This commit updates the Grab component to improve the handling of article URLs. It introduces the use of the "valid-url" package to validate URLs and sets the article URL based on the provided parameters. If the "url" parameter is present, it is used as the article URL. Otherwise, the code checks if the "text" or "title" parameters contain a valid URL and sets the article URL accordingly. If none of the parameters contain a valid URL, a message is displayed to the user indicating that the article cannot be saved. This change enhances the reliability and user experience of the Grab component.