refactor: Update serverExternalPackages in next.config.js
This commit updates the serverExternalPackages configuration in the next.config.js file. The commented out code block is replaced with a new block that includes the "puppeteer-extra", "puppeteer-extra-plugin-stealth", "puppeteer-extra-plugin-adblocker", "puppeteer-extra-plugin-block-resources", and "turndown" packages. This change ensures that the required server packages are included and improves the functionality of the application.